refactor to move getPolicyUrl into the deleteResource function
