+++ src/accessControl/acp/index.js
@@ -27,7 +27,6 @@ import {
   saveSolidDatasetAt,
   setThing,
 } from "@inrupt/solid-client";
-import { joinPath } from "../../stringHelpers";
 import { fetchProfile } from "../../solidClientHelpers/profile";
 import {
   ACL,
@@ -39,15 +38,9 @@ import {
   chainPromise,
   createResponder,
   isHTTPError,
-  sharedStart,
 } from "../../solidClientHelpers/utils";
 import { getOrCreateDataset } from "../../solidClientHelpers/resource";
-
-const POLICIES_CONTAINER = "pb_policies/";
-
-export function getPoliciesContainerUrl(podRootUri) {
-  return joinPath(podRootUri, POLICIES_CONTAINER);
-}
+import { getPolicyUrl } from "../../solidClientHelpers/policies";
 
 export function createAcpMap(read = false, write = false, append = false) {
   return {
@@ -85,18 +78,6 @@ export function getOrCreatePermission(permissions, webId) {
     access: createAcpMap(),
   };
   return permission;
-}
-
-export function getPolicyUrl(resource, policiesContainer) {
-  const resourceUrl = getSourceUrl(resource);
-  const policiesContainerUrl = getSourceUrl(policiesContainer);
-  const rootUrl = policiesContainerUrl.substr(
-    0,
-    policiesContainerUrl.length - POLICIES_CONTAINER.length
-  );
-  const matchingStart = sharedStart(resourceUrl, rootUrl);
-  const path = resourceUrl.substr(matchingStart.length);
-  return `${getPoliciesContainerUrl(matchingStart) + path}.ttl`;
 }
 
 export function getOrCreatePolicy(policyDataset, url) {
